+++ packages/transaction-controller/src/types.ts
@@ -4,11 +4,11 @@ import type { Operation } from 'fast-json-patch';
 /**
  * Representation of transaction metadata.
  */
-export type TransactionMeta =
-  | ({
-      status: Exclude<TransactionStatus, TransactionStatus.failed>;
-    } & TransactionMetaBase)
-  | ({ status: TransactionStatus.failed; error: Error } & TransactionMetaBase);
+export type TransactionMeta = TransactionMetaBase &
+  (
+    | { status: Exclude<TransactionStatus, TransactionStatus.failed> }
+    | { status: TransactionStatus.failed; error: Error }
+  );
 
 /**
  * Information about a single transaction such as status and block number.